Move Digester related classes
