+++ core/src/main/java/nl/nn/adapterframework/configuration/digester/AbstractSpringPoweredDigesterFactory.java
@@ -13,7 +13,7 @@
    See the License for the specific language governing permissions and
    limitations under the License.
 */
-package nl.nn.adapterframework.configuration;
+package nl.nn.adapterframework.configuration.digester;
 
 import java.beans.PropertyDescriptor;
 import java.lang.reflect.Method;
@@ -33,6 +33,11 @@ import org.xml.sax.Attributes;
 import org.xml.sax.Locator;
 
 import lombok.Setter;
+import nl.nn.adapterframework.configuration.ConfigurationException;
+import nl.nn.adapterframework.configuration.ConfigurationWarning;
+import nl.nn.adapterframework.configuration.ConfigurationWarnings;
+import nl.nn.adapterframework.configuration.HasSpecialDefaultValues;
+import nl.nn.adapterframework.configuration.SuppressKeys;
 import nl.nn.adapterframework.core.INamedObject;
 import nl.nn.adapterframework.util.LogUtil;
 import nl.nn.adapterframework.util.SpringUtils;
