+++ core/src/main/java/nl/nn/adapterframework/configuration/digester/JmsRealmsFactory.java
@@ -13,7 +13,7 @@
    See the License for the specific language governing permissions and
    limitations under the License.
 */
-package nl.nn.adapterframework.configuration;
+package nl.nn.adapterframework.configuration.digester;
 
 /**
  * Factory for the JMS Realms instance, which holds all JMS Realm definitions
